requirements.txt
# GUI Framework
PyQt5>=5.15.9
PyQt5-sip>=12.12.0

# DICOM Processing
pydicom>=2.4.4
pylibjpeg>=2.0.0
pylibjpeg-libjpeg>=2.0.0
pylibjpeg-openjpeg>=2.0.0
python-gdcm>=3.0.22

# Image Processing
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0

# JPEG 2000 Support
glymur>=0.12.0

# Encryption
cryptography>=41.0.0

# Utilities
python-dateutil>=2.8.0
